Split login and organization onboarding routes
